Various bug fixes
- When detecting the VM, check if the given path actually exists
instead of blindly assuming it does. Helps prevent issues when the path
was valid but was deleted.
- When cloning the VM, change the default site’s mapped path because
the user can’t be expected to have their site in a folder named Laravel.
- No longer allow reloading in the start function; .disabled class on
buttons does not actually disable the button, can be confusing if the
user for whatever reason clicks the start button when it appears to be
disabled and the VM is suddenly reloading.
- If the VM is being provisioned when it is not already running, use
`vagrant up` because `vagrant provision` does not work on stopped VMs,
also remove the _NEED_PROVISION state so that the stateChanged function
does not try to re-run the provisioning after starting the VM.